Split asset store implements into files
